[COLAB-2918]: separate service and controller in tracking service
