+++ microservices/services/tracking-service/src/main/java/org/xcolab/service/tracking/service/TrackedVisitService.java
@@ -5,14 +5,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import org.xcolab.client.tracking.ITrackingClient;
-import org.xcolab.client.tracking.pojo.ILocation;
 import org.xcolab.client.tracking.pojo.ITrackedVisit;
 import org.xcolab.client.tracking.pojo.ITrackedVisitor;
 import org.xcolab.service.tracking.domain.trackedVisitor.TrackedVisitorDao;
@@ -22,7 +16,7 @@ import org.xcolab.service.tracking.service.iptranslation.IpTranslationService.Ip
 
 @Service
 @RestController
-public class TrackedVisitService implements ITrackingClient {
+public class TrackedVisitService {
 
     private static final Logger log = LoggerFactory.getLogger(TrackedVisitService.class);
 
@@ -41,10 +35,7 @@ public class TrackedVisitService implements ITrackingClient {
         this.trackedVisitorDao = trackedVisitorDao;
     }
 
-    @Override
-    @RequestMapping(value = "/trackedVisits", method = RequestMethod.POST)
-    public ITrackedVisit addTrackedVisit(@RequestBody ITrackedVisit trackedVisit,
-            @RequestParam(required = false) Long userId) {
+    public ITrackedVisit createTrackedVisit(ITrackedVisit trackedVisit, Long userId) {
         final String remoteIp = trackedVisit.getIp();
         try {
             ipTranslationService.getLocationForIp(remoteIp).ifPresent(location -> {
@@ -70,18 +61,6 @@ public class TrackedVisitService implements ITrackingClient {
             trackedVisitorDao.update(trackedVisitor);
         }
         return trackedVisitDao.create(trackedVisit);
-    }
-
-    @Override
-    @RequestMapping(value = "/locations", method = RequestMethod.GET)
-    public ILocation getLocationForIp(@RequestParam String ipAddress) {
-        try {
-            return ipTranslationService.getLocationForIp(ipAddress)
-                    .orElse(null);
-        } catch (IpTranslationService.IpFormatException e) {
-            log.warn("Could not process ip address {}: {}", ipAddress, e.toString());
-            return null;
-        }
     }
 
     private boolean isLocalhost(String remoteIp) {
